try to batch UnityWebRequest
The Unity API for loading the audio files seems rather dumb, and
definitely not designed for local files. UnityWebRequest runs
in a background thread and takes a while, so waiting for each one
adds up if a mod has many small files. First do all requests, then
convert it to audio clips.